requirements.txt
connexion[swagger-ui]<3
flask>=2.0
setuptools==58.2.0
pycryptodome
Cmake  
youtube-dl
flask_session
pyrebase4
Werkzeug>=2.0
gunicorn>=20.0